Create README.md
In this Project we explore Covid 19 Data base. I uploaded the actual Covid 19 Data that was used in this Project. 

A visualization for this data was done using tableau and uploaded on Tableau Public on this link below:

https://public.tableau.com/app/profile/elsheikh7399/viz/CovidDashboardProject_16552040063230/Dashboard1

At First I split Covid Data into two files ( Covid death and Covid Vaccination). Both were uploaded and are accessible on GitHub. 


We Explored the following using Covid Death excel file:

- Total Cases vs Total Deaths
- Total Cases vs Population
- Countries with Highest Infection Rate compared to Population
- Countries with Highest Death Count per Population
- Continents with the highest death count per population
- Daily Total new Cases and Total Deaths

We then explored the amount the amount of people vaccinated in the world using Covid Vaccination Excel File.